Rename class to match controller and add tests
